+++ charproc.c
@@ -1,4 +1,4 @@
-/* $XTermId: charproc.c,v 1.1318 2013/11/23 12:59:49 tom Exp $ */
+/* $XTermId: charproc.c,v 1.1319 2013/11/26 20:38:11 tom Exp $ */
 
 /*
  * Copyright 1999-2012,2013 by Thomas E. Dickey
@@ -1469,7 +1469,7 @@ decode_scs(XtermWidget xw, int which, int prefix, int suffix)
     }
     if (result != nrc_Unknown) {
 	screen->gsets[which] = result;
-	TRACE(("setting G%d to %s\n", which, visibleScsCode(result)));
+	TRACE(("setting G%d to %s\n", which, visibleScsCode((int) result)));
     } else {
 	TRACE(("...unknown GSET\n"));
     }
